feat(web): add "About" page
