--- packages_rs/nextclade-web/src/components/About/About.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-import AboutContent from './AboutContent.mdx'
-
-export function About() {
-  return <AboutContent />
-}
