+++ packages_rs/nextclade-web/src/components/Layout/NavigationBar.tsx
@@ -12,6 +12,7 @@ import { Link } from 'src/components/Link/Link'
 import { FaDocker, FaGithub, FaXTwitter, FaDiscourse } from 'react-icons/fa6'
 import { LinkSmart } from 'src/components/Link/LinkSmart'
 import { ResultsStatus } from 'src/components/Results/ResultsStatus'
+import { PROJECT_NAME } from 'src/constants'
 import { canDownloadAtom, hasRanAtom, hasTreeAtom } from 'src/state/results.state'
 import styled, { useTheme } from 'styled-components'
 import { useTranslationSafe } from 'src/helpers/useTranslationSafe'
@@ -170,6 +171,11 @@ export function NavigationBar() {
         url: '/settings',
         content: t('Settings'),
         title: t('Configure Nextclade'),
+      },
+      {
+        url: '/about',
+        title: t('About {{what}}', { what: PROJECT_NAME }),
+        content: t('About'),
       },
       {
         title: t('Cite Nextclade in your work'),
