+++ packages_rs/nextclade-web/src/pages/about.tsx
@@ -1,0 +1,1 @@
+export { AboutPage as default } from 'src/components/About/AboutPage'
